covidmap/requirements.txt
requests
country-converter
pygal_maps_world
